Вынести работу с таблицей likes в LikeRepository. Удалить лишние классы исключений
