+++ src/main/java/ru/yandex/practicum/filmorate/exception/BadRequestException.java
@@ -1,0 +1,7 @@
+package ru.yandex.practicum.filmorate.exception;
+
+public class BadRequestException extends RuntimeException {
+    public BadRequestException(String message) {
+        super(message);
+    }
+}
