--- src/main/java/ru/yandex/practicum/filmorate/exception/FilmNotFoundException.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.filmorate.exception;
-
-public class FilmNotFoundException extends RuntimeException {
-    public FilmNotFoundException(String message) {
-        super(message);
-    }
-}
