--- src/main/java/ru/yandex/practicum/filmorate/exception/GenreNotFoundException.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.filmorate.exception;
-
-public class GenreNotFoundException extends RuntimeException {
-    public GenreNotFoundException(String message) {
-        super(message);
-    }
-}
