--- src/main/java/ru/yandex/practicum/filmorate/exception/MpaNotFoundException.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.filmorate.exception;
-
-public class MpaNotFoundException extends RuntimeException {
-    public MpaNotFoundException(String message) {
-        super(message);
-    }
-}
