--- src/main/java/ru/yandex/practicum/filmorate/exception/RequestException.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.filmorate.exception;
-
-public class RequestException extends RuntimeException {
-    public RequestException(String message) {
-        super(message);
-    }
-}
