--- src/main/java/ru/yandex/practicum/filmorate/exception/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.filmorate.exception;
-
-public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(String message) {
-        super(message);
-    }
-}
