+++ src/main/java/ru/yandex/practicum/filmorate/repository/film/FilmRepository.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.repository;
+package ru.yandex.practicum.filmorate.repository.film;
 
 import ru.yandex.practicum.filmorate.model.Film;
 
@@ -13,10 +13,6 @@ public interface FilmRepository {
     Collection<Film> getAllFilms();
 
     Film getFilmById(long id);
-
-    void addLike(long filmId, long userId);
-
-    void removeLike(long filmId, long userId);
 
     List<Film> getTopPopular(int count);
 }
