+++ src/main/java/ru/yandex/practicum/filmorate/repository/friend/FriendRepository.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.repository;
+package ru.yandex.practicum.filmorate.repository.friend;
 
 import java.util.List;
