+++ src/main/java/ru/yandex/practicum/filmorate/repository/like/LikeRepository.java
@@ -1,0 +1,7 @@
+package ru.yandex.practicum.filmorate.repository.like;
+
+public interface LikeRepository {
+    void addLike(long filmId, long userId);
+
+    void removeLike(long filmId, long userId);
+}
