+++ src/main/java/ru/yandex/practicum/filmorate/repository/mpa/MpaRepository.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.repository;
+package ru.yandex.practicum.filmorate.repository.mpa;
 
 import ru.yandex.practicum.filmorate.model.Mpa;
