+++ src/main/java/ru/yandex/practicum/filmorate/repository/user/UserRepository.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.repository;
+package ru.yandex.practicum.filmorate.repository.user;
 
 import ru.yandex.practicum.filmorate.model.User;
